+++ src/Umbraco.Web.Common/ApplicationBuilder/UmbracoApplicationBuilder.cs
@@ -1,4 +1,7 @@
-﻿using Microsoft.AspNetCore.Builder;
+using System;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Routing;
+using Umbraco.Cms.Core.Services;
 
 namespace Umbraco.Cms.Web.Common.ApplicationBuilder
 {
@@ -7,11 +10,17 @@ namespace Umbraco.Cms.Web.Common.ApplicationBuilder
     /// </summary>
     internal class UmbracoApplicationBuilder : IUmbracoApplicationBuilder
     {
-        public UmbracoApplicationBuilder(IApplicationBuilder appBuilder)
+        public UmbracoApplicationBuilder(IServiceProvider services, IRuntimeState runtimeState, IApplicationBuilder appBuilder, IEndpointRouteBuilder endpointRouteBuilder)
         {
+            ApplicationServices = services;
+            EndpointRouteBuilder = endpointRouteBuilder;
+            RuntimeState = runtimeState;
             AppBuilder = appBuilder;
         }
 
+        public IServiceProvider ApplicationServices { get; }
+        public IEndpointRouteBuilder EndpointRouteBuilder { get; }
+        public IRuntimeState RuntimeState { get; }
         public IApplicationBuilder AppBuilder { get; }
     }
 }
