+++ src/Umbraco.Web.Common/Extensions/ServiceCollectionExtensions.cs
@@ -16,7 +16,7 @@ using IHostingEnvironment = Umbraco.Cms.Core.Hosting.IHostingEnvironment;
 
 namespace Umbraco.Extensions
 {
-    public static class UmbracoCoreServiceCollectionExtensions
+    public static partial class ServiceCollectionExtensions
     {
         /// <summary>
         /// Create and configure the logger
