+++ src/Umbraco.Web.Website/DependencyInjection/UmbracoBuilderExtensions.cs
@@ -55,6 +55,8 @@ namespace Umbraco.Extensions
                 .AddDistributedCache()
                 .AddModelsBuilder();
 
+            builder.AddMembersIdentity();
+
             return builder;
         }
     }
